chore: move all dialogs to their own folder
