+++ packages/client/src/components/dialogs/BatteryInfoDialog.tsx
