+++ packages/client/src/components/dialogs/CastleHitDialog.tsx
@@ -1,3 +1,4 @@
+import { Button } from '@/components/ui/button';
 import {
   Dialog,
   DialogContent,
@@ -6,8 +7,6 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog';
 import { useBattle } from '@/contexts/BattleContext';
-
-import { Button } from './ui/button';
 
 export const CastleHitDialog: React.FC = () => {
   const {
