+++ packages/client/src/components/dialogs/HowToPlayDialog.tsx
@@ -1,4 +1,9 @@
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from './ui/dialog';
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from '@/components/ui/dialog';
 
 type HowToPlayDialogProps = {
   isHelpDialogOpen: boolean;
