+++ packages/client/src/components/dialogs/InstallInfoDialog.tsx
