+++ packages/client/src/pages/Battle.tsx
@@ -18,22 +18,22 @@ import { zeroAddress, zeroHash } from 'viem';
 import { useAccount } from 'wagmi';
 
 import { BackgroundAnimation } from '@/components/BackgroundAnimation';
-import { BatteryInfoDialog } from '@/components/BatteryInfoDialog';
 import { BattleBoard, INSTALLABLE_TOWERS } from '@/components/BattleBoard';
 import { BattleControlButtons } from '@/components/BattleControlButtons';
 import { BattleStatusBar } from '@/components/BattleStatusBar';
-import { CastleHitDialog } from '@/components/CastleHitDialog';
 import { ClickIndicator } from '@/components/ClickIndicator';
-import { ForfeitDialog } from '@/components/ForfeitDialog';
-import { HowToPlayDialog } from '@/components/HowToPlayDialog';
-import { InstallInfoDialog } from '@/components/InstallInfoDialog';
+import { BatteryInfoDialog } from '@/components/dialogs/BatteryInfoDialog';
+import { CastleHitDialog } from '@/components/dialogs/CastleHitDialog';
+import { ForfeitDialog } from '@/components/dialogs/ForfeitDialog';
+import { HowToPlayDialog } from '@/components/dialogs/HowToPlayDialog';
+import { InstallInfoDialog } from '@/components/dialogs/InstallInfoDialog';
+import { NoActionsDialog } from '@/components/dialogs/NoActionsDialog';
+import { PlayAgainDialog } from '@/components/dialogs/PlayAgainDialog';
+import { WelcomeDialog } from '@/components/dialogs/WelcomeDialog';
 import { LoadingScreen } from '@/components/LoadingScreen';
-import { NoActionsDialog } from '@/components/NoActionsDialog';
 import { NoBattleScreen } from '@/components/NoBattleScreen';
-import { PlayAgainDialog } from '@/components/PlayAgainDialog';
 import { TowerSelection } from '@/components/TowerSelection';
 import { Button } from '@/components/ui/button';
-import { WelcomeDialog } from '@/components/WelcomeDialog';
 import { BattleProvider, useBattle } from '@/contexts/BattleContext';
 import { useSettings } from '@/contexts/SettingsContext';
 import { useMUD } from '@/hooks/useMUD';
